+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog summarizes major changes between Graal versions relevant to langu
 ## `tip`
 ### Graal
 * Add experimental support constructing low-level IR in SSA form.
+* Add experimental support for SSA linear scan register allocation.
 ...
 
 ### Truffle